% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMetricsPlot.R
\name{createMetricsPlot}
\alias{createMetricsPlot}
\title{Create Metrics Plot}
\usage{
createMetricsPlot(POD = "North", target = 0.815, lowerBound = 0,
  graphtitle = "NTX Aetna Commercial - Generic Fill Rate (Primary Care)",
  graphname = paste0("C:/Users/e90795/Desktop/Rx1.png"))
}
\arguments{
\item{POD}{A character string. defaults to North}

\item{target}{Numeric. Goal for metric. Defaults to 0.815}

\item{lowerBound}{A character string. Lower bound of the graph. Defaults to 0.}

\item{graphtitle}{A character string.}

\item{graphname}{must end in .png.}
}
\description{
This function is creates a standard time series plot for the Joint Operation Committee (JOC) presenations
}
\examples{
createMetricsPlot()
}
\keyword{ggplot2}
\keyword{graph}

