% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_bswqa.R
\name{theme_bswqa}
\alias{theme_bswqa}
\title{BSWQA ggplot2 theme}
\usage{
theme_bswqa(base_size = 12, base_family = "sans")
}
\description{
Theme is standard used in BSWQA JOC presentations
}
\examples{
library(ggplot2)
g <- ggplot(mtcars) +
     geom_point(aes(x = wt, y = mpg, colour = factor(gear))) +
     theme_bswqa()
}
\keyword{ggplot2}
\keyword{graph}

